requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Data Processing and Augmentation
albumentations>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Configuration and Utilities
PyYAML>=6.0
typer>=0.9.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Model Export and Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
openvino>=2023.0.0

# Web API (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: TensorRT (for NVIDIA GPU optimization)
# tensorrt>=8.6.0  # Uncomment if you have TensorRT installed